created single test case for all options of page payout
