+++ tests/e2e/specs/customizer/global/container/cont-layout.test.js
@@ -1,17 +1,17 @@
 import { createURL } from '@wordpress/e2e-test-utils';
 import { setCustomize } from '../../../../utils/set-customize';
 describe( 'Container layout setting from Global option under the Customizer', () => {
-	it( 'Layout for container as BOXED should apply correctly', async () => {
-		const boxLayout = {
-			'site-content-layout': 'boxed-container',
+	it( 'Layout for container should apply correctly', async () => {
+		const contLayout = {
+			'single-page-content-layout': 'plain-container',
 		};
-		await setCustomize( boxLayout );
+		await setCustomize( contLayout );
 		await page.goto( createURL( 'sample-page' ), {
 			waitUntil: 'networkidle0',
 		} );
-		await page.waitForSelector('.site-content .ast-container');
+		await page.waitForSelector('.ast-container');
 		await expect( {
-			selector: '.site-content .ast-container',
+			selector: '.ast-container',
 			property: '',
 		} ).cssValueToBe( `` );
 	});
