+++ tests/e2e/specs/customizer/global/container/content-box.test.js
@@ -1,18 +1,19 @@
 import { createURL } from '@wordpress/e2e-test-utils';
 import { setCustomize } from '../../../../utils/set-customize';
 describe( 'Container layout setting from Global option under the Customizer', () => {
-	it( 'Layout for container should apply correctly', async () => {
-		const contLayout = {
-			'site-content-layout': 'plain-container',
+	it( 'Layout for container as CONTENT-BOXED should apply correctly', async () => {
+		const contboxLayout = {
+			'single-page-content-layout': 'content-boxed-container',
 		};
-		await setCustomize( contLayout );
+		await setCustomize( contboxLayout );
 		await page.goto( createURL( 'sample-page' ), {
 			waitUntil: 'networkidle0',
 		} );
-		await page.waitForSelector('.ast-container');
+		await page.waitForSelector('.site-content .ast-container');
 		await expect( {
-			selector: '.ast-container',
+			selector: '.site-content .ast-container',
 			property: '',
 		} ).cssValueToBe( `` );
 	});
 });
+//.ast-container max-width 1240px
