+++ tests/e2e/specs/customizer/global/container/stretch.test.js
@@ -3,7 +3,7 @@ import { setCustomize } from '../../../../utils/set-customize';
 describe( 'Container layout setting from Global option under the Customizer', () => {
 	it( 'Layout for container as FULL WIDTH/ STRETCH should apply correctly', async () => {
 		const contLayout = {
-			'site-content-layout': 'page-builder',
+			'single-page-content-layout': 'page-builder',
 		};
 		await setCustomize( contLayout );
 		await page.goto( createURL( 'sample-page' ), {
